+++ NEWS.md
@@ -8,6 +8,9 @@
 * `environments()` now has it's own `print.mark_environments()` method rather than calling `cat()` within the function itself
 * `array_extract()`'s first argument is changed from `arr` to `.arr`
 * `diff_time()` now defaults to using UTC (Related to [#32](https://github.com/jmbarbone/mark/issues/32))
+* `print.note()` method has been updated (Related to: [#33](https://github.com/jmbarbone/mark/issues/33)):
+  * to print `x` _normally_, without the `note class` when just the note has to be seen
+  * an internal function now handles the note formatted for class `noted`
 
 ## Fixes